lab9/userservice/requirements.txt
flask
pandas
numpy
sqlalchemy
sqlalchemy-utils
sqlalchemy-bigquery
pyjwt